Release the SIGABRT!
Changed NejvetsiNahVPoli.c to work with txt files
Added BlankLine.c which determines if string contains onlz whitespace
Added file.txt for testing purposes

STORY:
The unevitable human error occured, thus our prophet was forced to redo all the gooddoings and wrongdoing he ever ever did, which lead to even more evil in these far lands..

ISSUES:
There is something fishy going around with realloc... SIGABRT jumpscares
me like no tomorrow.

+++ Cvičení/IndexNejvetsiho.c
@@ -1,5 +1,5 @@
 #include "IndexNejvetsiho.h"
-unsigned int IndexNejvetsiho(int *arr, unsigned int arrlength){
+unsigned int IndexNejvetsiho(long int *arr, unsigned int arrlength){
         
         unsigned int largestpos = 0;
